chore: updating tests with todos
